+++ configs/prepare_dataset.yaml
@@ -9,7 +9,7 @@ YOLO:
   pose_ckpt: "/workspace/code/ckpt/yolo11x-pose.pt"
   seg_ckpt: "/workspace/code/ckpt/yolo11x-seg.pt"
   bbox_ckpt: "/workspace/code/ckpt/yolo11x.pt"
-  conf: 0.5
+  conf: 0.8
   iou: 0.5
   verbose: False
   img_size: 640
