+++ prepare_dataset/depth_estimation.py
@@ -28,6 +28,8 @@ import torch
 import numpy as np
 from PIL import Image
 
+from utils.utils import save_to_pt, merge_frame_to_video
+
 
 class DepthEstimator:
     def __init__(self, configs):
@@ -44,7 +46,7 @@ class DepthEstimator:
         ).to(self.device)
 
         self.save = configs.depth_estimator.save
-        self.save_path = Path(configs.extract_dataset.save_path) / "vis" / "depth"
+        self.save_path = Path(configs.extract_dataset.save_path)
 
     def estimate_depth(self, frames: np.ndarray):
 
@@ -75,7 +77,7 @@ class DepthEstimator:
         person = video_path.parts[-2]
         video_name = video_path.stem
 
-        _save_path = self.save_path / person / video_name
+        _save_path = self.save_path / "vis" / "depth" / person / video_name
         if not _save_path.exists():
             _save_path.mkdir(parents=True, exist_ok=True)
 
@@ -103,5 +105,10 @@ class DepthEstimator:
                 self.save_image(depths, i, video_path)
 
             res_depth.append(depths.cpu())
+
+        if self.save:
+            merge_frame_to_video(
+                self.save_path, video_path.parts[-2], video_path.stem, "depth"
+            )
 
         return torch.cat(res_depth, dim=0)
