+++ utils/__init__.py
@@ -1,14 +1,14 @@
-'''
+"""
 File: __init__.py
 Project: utils
 Created Date: 2023-10-20 08:17:12
 Author: chenkaixu
 -----
 Comment:
- 
+
 Have a good code time!
 -----
-Last Modified: Saturday April 6th 2024 8:08:31 am
+Last Modified: Monday May 5th 2025 9:08:50 pm
 Modified By: the developer formerly known as Kaixu Chen at <chenkaixusan@gmail.com>
 -----
 HISTORY:
@@ -17,6 +17,4 @@ Date 	By 	Comments
 
 26-11-2024	Kaixu Chen	refactor the code
 
-'''
-
-from project.utils.utils import *
+"""
